Separate LayoutPreferences and LayoutPrefsDb classes
